Create separate entity ApplicantOnVacancyStatus.
