+++ huntflow_api_client/entities/applicant_on_vacancy.py
@@ -1,4 +1,4 @@
-from huntflow_api_client.entities.base import BaseEntity, ListEntityMixin
+from huntflow_api_client.entities.base import BaseEntity
 from huntflow_api_client.models.request.applicant_on_vacancy import (
     AddApplicantToVacancyRequest,
     ApplicantVacancySplitRequest,
@@ -7,22 +7,10 @@ from huntflow_api_client.models.request.applicant_on_vacancy import (
 from huntflow_api_client.models.response.applicant_on_vacancy import (
     AddApplicantToVacancyResponse,
     ApplicantVacancySplitResponse,
-    VacancyStatusesResponse,
 )
 
 
-class ApplicantOnVacancy(BaseEntity, ListEntityMixin):
-    async def list(self, account_id: int) -> VacancyStatusesResponse:
-        """
-        API method reference
-            https://api.huntflow.ai/v2/docs#get-/accounts/-account_id-/vacancies/statuses
-
-        :param account_id: Organization ID
-        :return: List of available applicant on vacancy statuses (stages)
-        """
-        response = await self._api.request("GET", f"/accounts/{account_id}/vacancies/statuses")
-        return VacancyStatusesResponse.parse_obj(response.json())
-
+class ApplicantOnVacancy(BaseEntity):
     async def attach_applicant_to_vacancy(
         self,
         account_id: int,
